Svelte Handbook
Svelte is an exciting Web frontend framework that you can use to build Web
applications.
If you're just starting out, Svelte is a great choice as your first frontend
framework.
If you are already experienced in React, Vue.js, Angular or another frontend
frameworks you will be pleasantly surprised by Svelte.
Compared to React, Vue, Angular and other frameworks, an app built using
Svelte is compiled beforehand so you don't have to serve the whole
framework to every one of your site visitors.
As a result, the fruition of the experience is smoother, and consumes less
bandwidth and everything feels faster and more lightweight.
At deployment, Svelte disappears and all you get is plain (and fast!)
JavaScript.
This is just the tip of the iceberg. Let's get into it!